+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/room/timeline/element/message/FileBasedRoomMessageTimelineElementView.kt
@@ -90,9 +90,7 @@ class FileBasedRoomMessageTimelineElementViewImpl : FileBasedRoomMessageTimeline
         FileBasedRoomMessageTimelineElementMessageBubble(
             holder,
             element,
-            {
-                element.showSaveDialog()
-            },
+            { element.showSaveDialog() },
             isPreview,
             displayProgressOverElement,
             index,