+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/message/FileBasedRoomMessageTimelineElementViewModel.kt
@@ -1,7 +1,9 @@
 package de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.message
 
 import de.connect2x.lognity.api.logger.error
+import de.connect2x.trixnity.client.media.InsufficientSpaceException
 import de.connect2x.trixnity.client.media.PlatformMedia
+import de.connect2x.trixnity.clientserverapi.client.DownloadLimitExceededException
 import de.connect2x.trixnity.clientserverapi.model.media.FileTransferProgress
 import de.connect2x.trixnity.core.model.RoomId
 import de.connect2x.trixnity.core.model.events.m.room.RoomMessageEventContent
@@ -96,7 +98,14 @@ abstract class FileBasedRoomMessageTimelineElementViewModel<C : RoomMessageEvent
                                 maxSize = maxMediaSizeInMemory,
                             )
                     }
-                    .onFailure { _loadMediaError.value = i18n.mediaCouldNotBeRead() }
+                    .onFailure { error ->
+                        _loadMediaError.value =
+                            when (error) {
+                                is DownloadLimitExceededException -> i18n.mediaExceedsDownloadLimit()
+                                is InsufficientSpaceException -> i18n.mediaInsufficientSpace()
+                                else -> i18n.mediaCouldNotBeRead()
+                            }
+                    }
             }
             .invokeOnCompletion { activeLoadMedia.value = null }
     }
@@ -175,9 +184,13 @@ abstract class FileBasedRoomMessageTimelineElementViewModel<C : RoomMessageEvent
                             _downloadMedia.value = it
                             return Result.success(it)
                         },
-                        onFailure = {
-                            _downloadMediaError.value = i18n.downloadFailed(it.message)
-                            return Result.failure(it)
+                        onFailure = { error ->
+                            _downloadMediaError.value =
+                                when (error) {
+                                    is InsufficientSpaceException -> i18n.downloadFailedInsufficientSpace()
+                                    else -> i18n.downloadFailed(error.message)
+                                }
+                            return Result.failure(error)
                         },
                     )
             } catch (exc: CancellationException) {